Refactoring the `model` package (#239)
* Moar tests.
* Separate out Docker functions from the model package
* gofmt
* gofmt
* remove key operation examples for now
* Forgot to actually start the containers.
* Additional running check for container state searches. (#237)
* add dist to gitignore
* Update output path
* Decouple regression tests from using docker client api directly and use docker lib
* Add some notes on local development
* add service tests into travis
