+++ main_test.go
@@ -1,7 +1,6 @@
-package main
+package main_test
 
 import (
-	"context"
 	"fmt"
 	"os"
 	"testing"
@@ -10,13 +9,13 @@ import (
 	"github.com/docker/docker/api/types"
 	"github.com/docker/docker/api/types/container"
 	"github.com/docker/docker/api/types/network"
-	"github.com/docker/docker/client"
-	model "github.com/fubarhouse/pygmy-go/service/interface"
+	"github.com/fubarhouse/pygmy-go/service/interface/docker"
 	. "github.com/smartystreets/goconvey/convey"
 )
 
 const (
 	dindContainerName = "exampleTestContainer"
+	binaryReference   = "pygmy-go-linux-x86"
 )
 
 var (
@@ -37,54 +36,46 @@ type config struct {
 // run to keep the consistency for as many tests as are required.
 func setup(t *testing.T, config *config) {
 
-	var cleanCmd = "/builds/pygmy-go-linux-x86 clean"
-	var statusCmd = "/builds/pygmy-go-linux-x86 status"
-	var upCmd = "/builds/pygmy-go-linux-x86 up"
-	var downCmd = "/builds/pygmy-go-linux-x86 down"
+	var cleanCmd = fmt.Sprintf("/builds/%v clean", binaryReference)
+	var statusCmd = fmt.Sprintf("/builds/%v status", binaryReference)
+	var upCmd = fmt.Sprintf("/builds/%v up", binaryReference)
+	var downCmd = fmt.Sprintf("/builds/%v down", binaryReference)
 
 	if config.configpath != "" {
-		cleanCmd = fmt.Sprintf("/builds/pygmy-go-linux-x86 clean --config %v", config.configpath)
-		statusCmd = fmt.Sprintf("/builds/pygmy-go-linux-x86 status --config %v", config.configpath)
-		upCmd = fmt.Sprintf("/builds/pygmy-go-linux-x86 up --config %v", config.configpath)
-		downCmd = fmt.Sprintf("/builds/pygmy-go-linux-x86 dow --config %v", config.configpath)
+		cleanCmd = fmt.Sprintf("/builds/%v clean --config %v", binaryReference, config.configpath)
+		statusCmd = fmt.Sprintf("/builds/%v status --config %v", binaryReference, config.configpath)
+		upCmd = fmt.Sprintf("/builds/%v up --config %v", binaryReference, config.configpath)
+		downCmd = fmt.Sprintf("/builds/%v dow --config %v", binaryReference, config.configpath)
 	}
 
 	Convey("Pygmy Application Test: "+config.name, t, func() {
 
-		ctx := context.Background()
-		cli, err := client.NewClientWithOpts()
-		cli.NegotiateAPIVersion(ctx)
-
 		Convey("Provision environment", func() {
-			Convey("Connection to Docker Client", func() {
-				So(err, ShouldEqual, nil)
-			})
-
 			Convey("Image pulled", func() {
-				_, e := model.DockerPull("library/docker:dind")
+				_, e := docker.DockerPull("library/docker:dind")
 				So(e, ShouldBeNil)
 			})
 
 			Convey("Container created", func() {
 				currentWorkingDirectory, err := os.Getwd()
 				So(err, ShouldBeNil)
-				x, _ := cli.ContainerCreate(ctx, &container.Config{
+				x, _ := docker.DockerContainerCreate(dindContainerName, container.Config{
 					Image: "docker:dind",
-				}, &container.HostConfig{
+				}, container.HostConfig{
 					AutoRemove: false,
 					Binds: []string{
 						fmt.Sprintf("%v%vbuilds%v:/builds", currentWorkingDirectory, string(os.PathSeparator), string(os.PathSeparator)),
 						fmt.Sprintf("%v%vexamples%v:/examples", currentWorkingDirectory, string(os.PathSeparator), string(os.PathSeparator)),
 					},
 					Privileged: true,
-				}, &network.NetworkingConfig{}, dindContainerName)
+				}, network.NetworkingConfig{})
 
 				dindID = x.ID
 				So(dindID, ShouldNotEqual, "")
 			})
 
 			Convey("Container started", func() {
-				err = cli.ContainerStart(ctx, dindContainerName, types.ContainerStartOptions{})
+				err := docker.DockerContainerStart(dindContainerName, types.ContainerStartOptions{})
 				So(err, ShouldEqual, nil)
 			})
 		})
@@ -92,19 +83,19 @@ func setup(t *testing.T, config *config) {
 		Convey("Populating Daemon", func() {
 
 			Convey("Container has started the daemon", func() {
-				_, e := model.DockerExec(dindContainerName, "dockerd")
+				_, e := docker.DockerExec(dindContainerName, "dockerd")
 				So(e, ShouldEqual, nil)
 				time.Sleep(time.Second * 2)
 			})
 
-			e := cli.ContainerStart(ctx, dindContainerName, types.ContainerStartOptions{})
+			e := docker.DockerContainerStart(dindContainerName, types.ContainerStartOptions{})
 			if e != nil {
 				fmt.Println(e)
 			}
 
 			for _, image := range config.images {
 				Convey("Pulling "+image, func() {
-					_, e := model.DockerExec(dindContainerName, "docker pull "+image)
+					_, e := docker.DockerExec(dindContainerName, "docker pull "+image)
 					time.Sleep(time.Second * 2)
 					So(e, ShouldBeNil)
 				})
@@ -114,7 +105,7 @@ func setup(t *testing.T, config *config) {
 		Convey("Application Tests", func() {
 
 			Convey("Container has configuration file ("+config.configpath+")", func() {
-				d, _ := model.DockerExec(dindContainerName, "stat "+config.configpath)
+				d, _ := docker.DockerExec(dindContainerName, "stat "+config.configpath)
 				if config.configpath == "" {
 					SkipSo(string(d), ShouldContainSubstring, config.configpath)
 				} else {
@@ -123,30 +114,30 @@ func setup(t *testing.T, config *config) {
 			})
 
 			Convey("Container has compiled binary from host", func() {
-				d, _ := model.DockerExec(dindContainerName, "stat /builds/pygmy-go-linux-x86")
-				So(string(d), ShouldContainSubstring, "/builds/pygmy-go-linux-x86")
+				d, _ := docker.DockerExec(dindContainerName, fmt.Sprintf("stat /builds/%v", binaryReference))
+				So(string(d), ShouldContainSubstring, fmt.Sprintf("/builds/%v", binaryReference))
 			})
 
-			d, _ := model.DockerExec(dindContainerName, "/builds/pygmy-go-linux-x86")
+			d, _ := docker.DockerExec(dindContainerName, fmt.Sprintf("/builds/%v", binaryReference))
 			Convey("Container can run pygmy", func() {
 				So(string(d), ShouldContainSubstring, "local containers for local development")
 			})
 
 			// While it's safe, we should clean the environment.
-			_, e := model.DockerExec(dindContainerName, cleanCmd)
+			_, e := docker.DockerExec(dindContainerName, cleanCmd)
 			if e != nil {
 				fmt.Println(e)
 			}
 
 			Convey("Default ports are not allocated", func() {
-				g, _ := model.DockerExec(dindContainerName, statusCmd)
+				g, _ := docker.DockerExec(dindContainerName, statusCmd)
 				for _, service := range config.servicewithports {
 					So(string(g), ShouldContainSubstring, service+" is able to start")
 				}
 			})
 
 			Convey("Pygmy started", func() {
-				d, _ = model.DockerExec(dindContainerName, upCmd)
+				d, _ = docker.DockerExec(dindContainerName, upCmd)
 				if config.configpath != "" {
 					So(string(d), ShouldContainSubstring, "Using config file: "+config.configpath)
 				}
@@ -156,7 +147,7 @@ func setup(t *testing.T, config *config) {
 			})
 
 			Convey("Endpoints are serving", func() {
-				d, _ = model.DockerExec(dindContainerName, statusCmd)
+				d, _ = docker.DockerExec(dindContainerName, statusCmd)
 				for _, endpoint := range config.endpoints {
 					if config.skipendpointchecks {
 						SkipSo(string(d), ShouldNotContainSubstring, "! "+endpoint)
@@ -170,11 +161,11 @@ func setup(t *testing.T, config *config) {
 		Convey("Environment Cleanup", func() {
 			Convey("Pygmy has cleaned the environment", func() {
 
-				_, e := model.DockerExec(dindContainerName, downCmd)
+				_, e := docker.DockerExec(dindContainerName, downCmd)
 				So(e, ShouldBeNil)
-				_, e = model.DockerExec(dindContainerName, cleanCmd)
+				_, e = docker.DockerExec(dindContainerName, cleanCmd)
 				So(e, ShouldBeNil)
-				d, _ := model.DockerExec(dindContainerName, statusCmd)
+				d, _ := docker.DockerExec(dindContainerName, statusCmd)
 				for _, service := range config.services {
 					So(string(d), ShouldContainSubstring, service+" is not running")
 				}
@@ -182,13 +173,9 @@ func setup(t *testing.T, config *config) {
 			})
 			// System prune container...
 			Convey("Removing DinD Container", func() {
-				ctx := context.Background()
-				cli, err := client.NewClientWithOpts()
-				cli.NegotiateAPIVersion(ctx)
+				err := docker.DockerKill("exampleTestContainer")
 				So(err, ShouldBeNil)
-				err = cli.ContainerKill(ctx, "exampleTestContainer", "")
-				So(err, ShouldBeNil)
-				err = cli.ContainerRemove(ctx, "exampleTestContainer", types.ContainerRemoveOptions{Force: true})
+				err = docker.DockerRemove("exampleTestContainer")
 				So(err, ShouldBeNil)
 			})
 		})
@@ -204,7 +191,7 @@ func TestDefault(t *testing.T) {
 		images:             []string{"amazeeio/haproxy", "andyshinn/dnsmasq:2.78", "mailhog/mailhog"},
 		services:           []string{"amazeeio-haproxy", "amazeeio-dnsmasq", "amazeeio-mailhog"},
 		servicewithports:   []string{"amazeeio-haproxy", "amazeeio-mailhog"},
-		skipendpointchecks: true,
+		skipendpointchecks: false,
 	}
 	setup(t, configuration)
 }
