+++ service/interface/interface.go
@@ -1,60 +1,14 @@
 package model
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"io"
-	"io/ioutil"
-	"regexp"
-	"strconv"
 	"strings"
-	"time"
 
 	"github.com/docker/docker/api/types"
-	"github.com/docker/docker/api/types/container"
-	"github.com/docker/docker/api/types/filters"
-	"github.com/docker/docker/api/types/network"
-	volumetypes "github.com/docker/docker/api/types/volume"
 	"github.com/docker/docker/client"
-	"github.com/fubarhouse/pygmy-go/service/endpoint"
+	"github.com/fubarhouse/pygmy-go/service/interface/docker"
 )
-
-// DockerService is the requirements for a Docker container to be compatible.
-// The Service struct is used to implement this interface, and individual
-// variables of type Service can/have overwritten them when logic deems
-// it necessary.
-type DockerService interface {
-	Setup() error
-	Status() (bool, error)
-	Start() ([]byte, error)
-	Stop() error
-}
-
-// Service is a collection of requirements for starting a container and
-// provides a way for config of any container to be overridden and start
-// fully compatible with Docker's API.
-type Service struct {
-	Config        container.Config
-	HostConfig    container.HostConfig
-	NetworkConfig network.NetworkingConfig
-}
-
-// Network is a struct containing the configuration of a single Docker network
-// including some extra fields so that Pygmy knows how to interact with the
-// desired outcome.
-type Network struct {
-	// Name is the name of the network, it is independent of the map key which
-	// will be used to configure pygmy but this field should match the map key.
-	Name string `yaml:"name"`
-	// Containers is a []string which indicates the names of the containers
-	// that need to be connected to this network.
-	Containers []string `yaml:"containers"`
-	// Config is the actual Network configuration for the Docker Network.
-	// It is the Network creation configuration as provided by the Docker API.
-	Config types.NetworkCreate `yaml:"config"`
-}
 
 // Setup will detect if the Service's image reference exists and will
 // attempt to run `docker pull` on the non-canonical image if it is
@@ -64,14 +18,14 @@ func (Service *Service) Setup() error {
 		return nil
 	}
 
-	images, _ := DockerImageList()
+	images, _ := docker.DockerImageList()
 	for _, image := range images {
 		if strings.Contains(fmt.Sprint(image.RepoTags), Service.Config.Image) {
 			return nil
 		}
 	}
 
-	msg, err := DockerPull(Service.Config.Image)
+	msg, err := docker.DockerPull(Service.Config.Image)
 	if msg != "" {
 		fmt.Println(msg)
 	}
@@ -112,21 +66,21 @@ func (Service *Service) Start() ([]byte, error) {
 	}
 
 	if purpose == "addkeys" || purpose == "showkeys" {
-		if e := DockerKill(name); e != nil {
+		if e := docker.DockerKill(name); e != nil {
 			fmt.Sprintln(e)
 		}
-		if e := DockerRemove(name); e != nil {
+		if e := docker.DockerRemove(name); e != nil {
 			fmt.Sprintln(e)
 		}
 
 	}
 
-	output, err := DockerRun(Service)
+	output, err := Service.DockerRun()
 	if err != nil {
 		return []byte{}, err
 	}
 
-	if c, err := GetRunning(Service); c.ID != "" {
+	if c, err := Service.GetRunning(); c.ID != "" {
 		if !Service.HostConfig.AutoRemove && !discrete {
 			fmt.Printf("Successfully started %v\n", name)
 		} else if Service.HostConfig.AutoRemove && err != nil {
@@ -149,15 +103,7 @@ func (Service *Service) Status() (bool, error) {
 	if Service.HostConfig.AutoRemove {
 		return true, nil
 	}
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		fmt.Println(err)
-	}
-	containers, _ := cli.ContainerList(ctx, types.ContainerListOptions{
-		Quiet: true,
-	})
+	containers, _ := docker.DockerContainerList()
 	for _, container := range containers {
 		for _, n := range container.Names {
 			if strings.Contains(n, name) && strings.HasPrefix(container.Status, "Up") {
@@ -170,113 +116,10 @@ func (Service *Service) Status() (bool, error) {
 
 }
 
-// SetField will set a pygmy label to be equal to the string equal of
-// an interface{}, even if it already exists. It should not matter if
-// this container is running or not.
-func (Service *Service) SetField(name string, value interface{}) error {
-	if _, ok := Service.Config.Labels["pygmy."+fmt.Sprint(name)]; !ok {
-		//
-	} else {
-		old, _ := Service.GetFieldString(name)
-		Service.Config.Labels["pygmy."+name] = fmt.Sprint(value)
-		new, _ := Service.GetFieldString(name)
-
-		if old == new {
-			return fmt.Errorf("tag was not set")
-		}
-	}
-
-	return nil
-}
-
-// GetFieldString will get and return a tag on the service using the pygmy
-// convention ("pygmy.*") and return it as a string.
-func (Service *Service) GetFieldString(field string) (string, error) {
-
-	f := fmt.Sprintf("pygmy.%v", field)
-
-	if container, running := GetRunning(Service); running == nil {
-		if val, ok := container.Labels[f]; ok {
-			return val, nil
-		}
-	}
-
-	if val, ok := Service.Config.Labels[f]; ok {
-		return val, nil
-	}
-
-	return "", fmt.Errorf("could not find field 'pygmy.%v' on service using image %v?", field, Service.Config.Image)
-}
-
-// GetFieldInt will get and return a tag on the service using the pygmy
-// convention ("pygmy.*") and return it as an int.
-func (Service *Service) GetFieldInt(field string) (int, error) {
-
-	f := fmt.Sprintf("pygmy.%v", field)
-
-	if container, running := GetRunning(Service); running == nil {
-		if val, ok := container.Labels[f]; ok {
-			i, e := strconv.ParseInt(val, 10, 10)
-			if e != nil {
-				return 0, e
-			}
-			return int(i), nil
-		}
-	}
-
-	if val, ok := Service.Config.Labels[f]; ok {
-		i, e := strconv.ParseInt(val, 10, 10)
-		if e != nil {
-			return 0, e
-		}
-		return int(i), nil
-	}
-
-	return 0, fmt.Errorf("could not find field 'pygmy.%v' on service using image %v?", field, Service.Config.Image)
-}
-
-// GetFieldBool will get and return a tag on the service using the pygmy
-// convention ("pygmy.*") and return it as a bool.
-func (Service *Service) GetFieldBool(field string) (bool, error) {
-
-	f := fmt.Sprintf("pygmy.%v", field)
-
-	if container, running := GetRunning(Service); running == nil {
-		if Service.Config.Labels[f] == container.Labels[f] {
-			if val, ok := container.Labels[f]; ok {
-				if val == "true" {
-					return true, nil
-				} else if val == "false" {
-					return false, nil
-				}
-			}
-		}
-	}
-
-	if val, ok := Service.Config.Labels[f]; ok {
-		if val == "true" || val == "1" {
-			return true, nil
-		} else if val == "false" || val == "0" {
-			return false, nil
-		}
-	}
-
-	return false, fmt.Errorf("could not find field 'pygmy.%v' on service using image %v?", field, Service.Config.Image)
-}
-
 // GetRunning will get a types.Container variable for a given running container
 // and it will not retrieve any information on containers that are not running.
-func GetRunning(Service *Service) (types.Container, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		fmt.Println(err)
-	}
-	containers, _ := cli.ContainerList(ctx, types.ContainerListOptions{
-		Quiet: true,
-	})
-
+func (Service *Service) GetRunning() (types.Container, error) {
+	containers, _ := docker.DockerContainerList()
 	for _, container := range containers {
 		if _, ok := container.Labels["pygmy.name"]; ok {
 			if strings.Contains(container.Names[0], Service.Config.Labels["pygmy.name"]) {
@@ -296,22 +139,22 @@ func (Service *Service) Clean() error {
 		return nil
 	}
 
-	Containers, _ := DockerContainerList()
+	Containers, _ := docker.DockerContainerList()
 	for _, container := range Containers {
 		if container.Names[0] == name {
 			if pygmy {
 				name := strings.TrimLeft(container.Names[0], "/")
-				if e := DockerKill(container.ID); e == nil {
+				if e := docker.DockerKill(container.ID); e == nil {
 					if !Service.HostConfig.AutoRemove {
 						fmt.Printf("Successfully killed %v\n", name)
 					}
 				}
-				if e := DockerStop(container.ID); e == nil {
+				if e := docker.DockerStop(container.ID); e == nil {
 					if !Service.HostConfig.AutoRemove {
 						fmt.Printf("Successfully stopped %v\n", name)
 					}
 				}
-				if e := DockerRemove(container.ID); e != nil {
+				if e := docker.DockerRemove(container.ID); e != nil {
 					if !Service.HostConfig.AutoRemove {
 						fmt.Printf("Successfully removed %v\n", name)
 					}
@@ -332,7 +175,7 @@ func (Service *Service) Stop() error {
 		return nil
 	}
 
-	container, err := GetRunning(Service)
+	container, err := Service.GetRunning()
 	if err != nil {
 		if !discrete {
 			fmt.Printf("Not running %v\n", name)
@@ -341,8 +184,8 @@ func (Service *Service) Stop() error {
 	}
 
 	for _, name := range container.Names {
-		if e := DockerStop(container.ID); e == nil {
-			if e := DockerRemove(container.ID); e == nil {
+		if e := docker.DockerStop(container.ID); e == nil {
+			if e := docker.DockerRemove(container.ID); e == nil {
 				if !discrete {
 					containerName := strings.Trim(name, "/")
 					fmt.Printf("Successfully removed %v\n", containerName)
@@ -357,149 +200,8 @@ func (Service *Service) Stop() error {
 // _ will ensure DockerService is implemented by Service.
 var _ DockerService = (*Service)(nil)
 
-// DockerContainerList will return a slice of containers
-func DockerContainerList() ([]types.Container, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	containers, err := cli.ContainerList(ctx, types.ContainerListOptions{
-		All: true,
-	})
-	if err != nil {
-		return []types.Container{}, err
-	}
-
-	return containers, nil
-
-}
-
-// DockerImageList will return a slice of Docker images.
-func DockerImageList() ([]types.ImageSummary, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	images, err := cli.ImageList(ctx, types.ImageListOptions{})
-	if err != nil {
-		return []types.ImageSummary{}, err
-	}
-
-	return images, nil
-
-}
-
-// DockerPull will pull a Docker image into the daemon.
-func DockerPull(image string) (string, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	{
-
-		// To support image references from external sources to docker.io we need to check
-		// and validate the image reference for all known cases of validity.
-
-		if m, _ := regexp.MatchString("^([a-zA-Z0-9]+[/][a-zA-Z0-9:-_]+[a-zA-Z0-9:-_.]+)$", image); m {
-			// URL was not provided (in full), but the tag was provided.
-			// For this, we prepend 'docker.io/' to the reference.
-			// Examples:
-			//  - amazeeio/pygmy:latest
-			image = fmt.Sprintf("docker.io/%v", image)
-		} else if m, _ := regexp.MatchString("^([a-zA-Z0-9]+[/][a-zA-Z0-9:-]+[a-zA-Z0-9:-_.]+)$", image); m {
-			// URL was not provided (in full), but the tag was not provided.
-			// For this, we prepend 'docker.io/' to the reference.
-			// Examples:
-			//  - amazeeio/pygmy
-			image = fmt.Sprintf("docker.io/%v", image)
-		} else if m, _ := regexp.MatchString("^([a-zA-Z0-9.].+[a-zA-Z0-9]+[/][a-zA-Z0-9:-_]+[a-zA-Z0-9:-_.]+)$", image); m {
-			// URL was provided (in full), but the tag was provided.
-			// For this, we do not alter the value provided.
-			// Examples:
-			//  - quay.io/amazeeio/pygmy:latest
-		} else if m, _ := regexp.MatchString("^([a-zA-Z0-9.].+[a-zA-Z0-9]+[/][a-zA-Z0-9:-_]+)$", image); m {
-			// URL was provided (in full), but the tag was not provided.
-			// For this, we do not alter the value provided.
-			// Examples:
-			//  - quay.io/amazeeio/pygmy
-		} else if m, _ := regexp.MatchString("^([a-zA-Z0-9]+[:][a-zA-Z0-9.-_]+)$", image); m {
-			// Library image was provided with tag identifier.
-			// For this, we prepend 'docker.io/' to the reference.
-			// Examples:
-			//  - pygmy:latest
-			image = fmt.Sprintf("docker.io/%v", image)
-		} else if m, _ := regexp.MatchString("^([a-zA-Z0-9-_]+)$", image); m {
-			// Library image was provided without tag identifier.
-			// For this, we prepend 'docker.io/' to the reference.
-			// Examples:
-			//  - pygmy
-			image = fmt.Sprintf("docker.io/%v", image)
-		} else {
-			// Validation not successful
-			return "", fmt.Errorf("error: regexp validation for %v failed", image)
-		}
-	}
-
-	// DockerHub Registry causes a stack trace fatal error when unavailable.
-	// We can check for this and report back, handling it gracefully and
-	// tell the user the service is down momentarily, and to try again shortly.
-	if strings.HasPrefix(image, "docker.io") {
-		if s := endpoint.Validate("https://registry-1.docker.io"); !s {
-			return "", fmt.Errorf("cannot reach the Docker Hub Registry, please try again in a few minutes.")
-		}
-	}
-
-	data, err := cli.ImagePull(ctx, image, types.ImagePullOptions{})
-	if err != nil {
-		fmt.Println(err)
-	}
-
-	d := json.NewDecoder(data)
-
-	type Event struct {
-		Status         string `json:"status"`
-		Error          string `json:"error"`
-		Progress       string `json:"progress"`
-		ProgressDetail struct {
-			Current int `json:"current"`
-			Total   int `json:"total"`
-		} `json:"progressDetail"`
-	}
-
-	var event *Event
-	for {
-		if err := d.Decode(&event); err != nil {
-			if err == io.EOF {
-				break
-			}
-
-			panic(err)
-		}
-	}
-
-	if event != nil {
-		if strings.Contains(event.Status, fmt.Sprint("Downloaded newer image")) {
-			return fmt.Sprintf("Successfully pulled %v", image), nil
-		}
-
-		if strings.Contains(event.Status, fmt.Sprint("Image is up to date")) {
-			return fmt.Sprintf("Image %v is up to date", image), nil
-		}
-	}
-	return event.Status, nil
-}
-
 // DockerRun will setup and run a given container.
-func DockerRun(Service *Service) ([]byte, error) {
+func (Service *Service) DockerRun() ([]byte, error) {
 
 	ctx := context.Background()
 	cli, err := client.NewClientWithOpts()
@@ -509,7 +211,7 @@ func DockerRun(Service *Service) ([]byte, error) {
 	}
 
 	// Ensure we have the image available:
-	images, _ := DockerImageList()
+	images, _ := docker.DockerImageList()
 
 	// Specify a false boolean which we can switch to true if the image is in the registry:
 	imageFound := false
@@ -535,7 +237,7 @@ func DockerRun(Service *Service) ([]byte, error) {
 	}
 
 	// Sanity check to ensure we don't get name conflicts.
-	c, _ := DockerContainerList()
+	c, _ := docker.DockerContainerList()
 	for _, cn := range c {
 		if strings.HasSuffix(cn.Names[0], Service.Config.Labels["pygmy.name"]) {
 			return []byte{}, nil
@@ -548,286 +250,15 @@ func DockerRun(Service *Service) ([]byte, error) {
 		return []byte{}, fmt.Errorf("container config is missing label for name")
 	}
 
-	resp, err := cli.ContainerCreate(ctx, &Service.Config, &Service.HostConfig, &Service.NetworkConfig, name)
+	resp, err := docker.DockerContainerCreate(name, Service.Config, Service.HostConfig, Service.NetworkConfig)
 	if err != nil {
 		return []byte{}, err
 	}
 
-	if err := cli.ContainerStart(ctx, resp.ID, types.ContainerStartOptions{}); err != nil {
+	if err := docker.DockerContainerStart(name, types.ContainerStartOptions{}); err != nil {
 		return []byte{}, err
 	}
 
-	b, _ := cli.ContainerLogs(ctx, resp.ID, types.ContainerLogsOptions{
-		ShowStdout: true,
-		ShowStderr: true,
-	})
+	return docker.DockerContainerLogs(resp.ID)
 
-	buf := new(bytes.Buffer)
-	if _, f := buf.ReadFrom(b); f != nil {
-		fmt.Println(f)
-	}
-
-	b.Close()
-
-	return buf.Bytes(), nil
-}
-
-// DockerStop will stop the container.
-func DockerStop(name string) error {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return err
-	}
-	timeout := time.Duration(10)
-	err = cli.ContainerStop(ctx, name, &timeout)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// DockerKill will kill the container.
-func DockerKill(name string) error {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return err
-	}
-	err = cli.ContainerKill(ctx, name, "")
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// DockerRemove will remove the container.
-// It will not remove the image.
-func DockerRemove(id string) error {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return err
-	}
-	err = cli.ContainerRemove(ctx, id, types.ContainerRemoveOptions{})
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// DockerNetworkCreate is an abstraction layer on top of the Docker API call
-// which will create a Docker network using a specified configuration.
-func DockerNetworkCreate(network *types.NetworkResource) error {
-	netVal, _ := DockerNetworkStatus(network.Name)
-	if netVal {
-		return fmt.Errorf("docker network %v already exists", network.Name)
-	}
-
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return err
-	}
-
-	config := types.NetworkCreate{
-		Driver:     network.Driver,
-		EnableIPv6: network.EnableIPv6,
-		IPAM:       &network.IPAM,
-		Internal:   network.Internal,
-		Attachable: network.Attachable,
-		Options:    network.Options,
-		Labels:     network.Labels,
-	}
-	_, err = cli.NetworkCreate(ctx, network.Name, config)
-	if err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// DockerNetworkRemove will attempt to remove a Docker network
-// and will not apply force to removal.
-func DockerNetworkRemove(network string) error {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return err
-	}
-	err = cli.NetworkRemove(ctx, network)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// DockerNetworkStatus will identify if a network with a
-// specified name is present been created and return a boolean.
-func DockerNetworkStatus(network string) (bool, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return false, err
-	}
-
-	networks, err := cli.NetworkList(ctx, types.NetworkListOptions{})
-	if err != nil {
-		return false, err
-	}
-
-	for _, n := range networks {
-		if n.Name == network {
-			return true, nil
-		}
-	}
-
-	return false, nil
-}
-
-// DockerNetworkGet will use the Docker API to retrieve a Docker network
-// which has a given name.
-func DockerNetworkGet(name string) (types.NetworkResource, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return types.NetworkResource{}, err
-	}
-	networks, err := cli.NetworkList(ctx, types.NetworkListOptions{})
-	if err != nil {
-		return types.NetworkResource{}, err
-	}
-	for _, network := range networks {
-		if val, ok := network.Labels["pygmy.name"]; ok {
-			if val == name {
-				return network, nil
-			}
-		}
-	}
-	return types.NetworkResource{}, nil
-}
-
-// DockerNetworkConnect will connect a container to a network.
-func DockerNetworkConnect(network string, containerName string) error {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return err
-	}
-	e := cli.NetworkConnect(ctx, network, containerName, nil)
-	if e != nil {
-		return e
-	}
-	return nil
-}
-
-// DockerNetworkConnect will check if a container is connected to a network.
-func DockerNetworkConnected(network string, containerName string) (bool, error) {
-	// Reset network state:
-	c, _ := DockerContainerList()
-	for d := range c {
-		if c[d].Labels["pygmy.name"] == containerName {
-			for net := range c[d].NetworkSettings.Networks {
-				if net == network {
-					return true, nil
-				}
-			}
-		}
-	}
-	return false, fmt.Errorf("network was found without the container connected")
-}
-
-// DockerVolumeExists will check if a Docker volume has been created.
-func DockerVolumeExists(volume types.Volume) (bool, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return false, err
-	}
-	_, _, err = cli.VolumeInspectWithRaw(ctx, volume.Name)
-	if err != nil {
-		return false, err
-	}
-
-	return true, nil
-}
-
-// DockerVolumeGet will return the full contents of a types.Volume from the API.
-func DockerVolumeGet(name string) (types.Volume, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-
-	if err != nil {
-		return types.Volume{
-			Name: name,
-		}, err
-	}
-
-	volumes, err := cli.VolumeList(ctx, filters.Args{})
-	if err != nil {
-		return types.Volume{
-			Name: name,
-		}, err
-	}
-
-	for _, volume := range volumes.Volumes {
-		if volume.Name == name {
-			return *volume, nil
-		}
-	}
-
-	return types.Volume{
-		Name: name,
-	}, nil
-}
-
-// DockerVolumeCreate will create a Docker Volume as configured.
-func DockerVolumeCreate(volume types.Volume) (types.Volume, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return types.Volume{}, err
-	}
-	return cli.VolumeCreate(ctx, volumetypes.VolumeCreateBody{
-		Driver:     volume.Driver,
-		DriverOpts: volume.Options,
-		Labels:     volume.Labels,
-		Name:       volume.Name,
-	})
-}
-
-// DockerExec will run a command in a Docker container and return the output.
-func DockerExec(container string, command string) ([]byte, error) {
-	ctx := context.Background()
-	cli, err := client.NewClientWithOpts()
-	cli.NegotiateAPIVersion(ctx)
-	if err != nil {
-		return []byte{}, err
-	}
-
-	if rst, err := cli.ContainerExecCreate(ctx, container, types.ExecConfig{
-		AttachStdout: true,
-		AttachStderr: true,
-		Cmd:          strings.Split(command, " ")}); err != nil {
-		return []byte{}, err
-	} else {
-		if response, err := cli.ContainerExecAttach(context.Background(), rst.ID, types.ExecStartCheck{}); err != nil {
-			return []byte{}, err
-		} else {
-			data, _ := ioutil.ReadAll(response.Reader)
-			defer response.Close()
-			return data, nil
-		}
-	}
 }
