+++ service/library/setup.go
@@ -11,6 +11,7 @@ import (
 	"github.com/fubarhouse/pygmy-go/service/dnsmasq"
 	"github.com/fubarhouse/pygmy-go/service/haproxy"
 	model "github.com/fubarhouse/pygmy-go/service/interface"
+	"github.com/fubarhouse/pygmy-go/service/interface/docker"
 	"github.com/fubarhouse/pygmy-go/service/mailhog"
 	"github.com/fubarhouse/pygmy-go/service/network"
 	"github.com/fubarhouse/pygmy-go/service/resolv"
@@ -155,7 +156,7 @@ func Setup(c *Config) {
 
 		for _, v := range c.Volumes {
 			// Get the potentially existing volume:
-			c.Volumes[v.Name], _ = model.DockerVolumeGet(v.Name)
+			c.Volumes[v.Name], _ = docker.DockerVolumeGet(v.Name)
 			// Merge the volume with the provided configuration:
 			c.Volumes[v.Name] = getVolume(c.Volumes[v.Name], c.Volumes[v.Name])
 		}
