+++ service/ssh/key/ssh_addkey_win.go
@@ -17,6 +17,7 @@ func NewAdder() model.Service {
 				"pygmy.defaults": "true",
 				"pygmy.enable":   "true",
 				"pygmy.name":     "amazeeio-ssh-agent-add-key",
+				"pygmy.network":  "amazeeio-network",
 				"pygmy.discrete": "true",
 				"pygmy.output":   "false",
 				"pygmy.purpose":  "addkeys",
@@ -25,33 +26,6 @@ func NewAdder() model.Service {
 		},
 		HostConfig: container.HostConfig{
 			IpcMode:     "private",
-			AutoRemove:  true,
-			VolumesFrom: []string{"amazeeio-ssh-agent"},
-		},
-		NetworkConfig: network.NetworkingConfig{},
-	}
-}
-
-// NewShower will provide the standard object for the SSH key shower container.
-func NewShower() model.Service {
-	return model.Service{
-		Config: container.Config{
-			Image: "amazeeio/ssh-agent",
-			Cmd: []string{
-				"ssh-add",
-				"-L",
-			},
-			Labels: map[string]string{
-				"pygmy.defaults": "true",
-				"pygmy.enable":   "true",
-				"pygmy.name":     "amazeeio-ssh-agent-show-keys",
-				"pygmy.discrete": "true",
-				"pygmy.output":   "false",
-				"pygmy.purpose":  "showkeys",
-				"pygmy.weight":   "32",
-			},
-		},
-		HostConfig: container.HostConfig{
 			AutoRemove:  true,
 			VolumesFrom: []string{"amazeeio-ssh-agent"},
 		},
