+++ service/ssh/key/ssh_showkeys.go
@@ -8,25 +8,29 @@ import (
 	model "github.com/fubarhouse/pygmy-go/service/interface"
 )
 
-// NewAdder will provide the standard object for the SSH key adder container.
-func NewAdder() model.Service {
+// NewShower will provide the standard object for the SSH key shower container.
+func NewShower() model.Service {
 	return model.Service{
 		Config: container.Config{
 			Image: "amazeeio/ssh-agent",
+			Cmd: []string{
+				"ssh-add",
+				"-L",
+			},
 			Labels: map[string]string{
 				"pygmy.defaults": "true",
 				"pygmy.enable":   "true",
-				"pygmy.name":     "amazeeio-ssh-agent-add-key",
+				"pygmy.name":     "amazeeio-ssh-agent-show-keys",
 				"pygmy.network":  "amazeeio-network",
 				"pygmy.discrete": "true",
 				"pygmy.output":   "false",
-				"pygmy.purpose":  "addkeys",
-				"pygmy.weight":   "31",
+				"pygmy.purpose":  "showkeys",
+				"pygmy.weight":   "32",
 			},
 		},
 		HostConfig: container.HostConfig{
-			IpcMode:     "private",
 			AutoRemove:  true,
+			IpcMode:     "private",
 			VolumesFrom: []string{"amazeeio-ssh-agent"},
 		},
 		NetworkConfig: network.NetworkingConfig{},
