+++ service/ssh/key/ssh_showkeys_win.go
@@ -1,4 +1,4 @@
-// +build darwin linux
+// +build windows
 
 package key
 
@@ -8,24 +8,26 @@ import (
 	model "github.com/fubarhouse/pygmy-go/service/interface"
 )
 
-// NewAdder will provide the standard object for the SSH key adder container.
-func NewAdder() model.Service {
+// NewShower will provide the standard object for the SSH key shower container.
+func NewShower() model.Service {
 	return model.Service{
 		Config: container.Config{
 			Image: "amazeeio/ssh-agent",
+			Cmd: []string{
+				"ssh-add",
+				"-L",
+			},
 			Labels: map[string]string{
 				"pygmy.defaults": "true",
 				"pygmy.enable":   "true",
-				"pygmy.name":     "amazeeio-ssh-agent-add-key",
-				"pygmy.network":  "amazeeio-network",
+				"pygmy.name":     "amazeeio-ssh-agent-show-keys",
 				"pygmy.discrete": "true",
 				"pygmy.output":   "false",
-				"pygmy.purpose":  "addkeys",
-				"pygmy.weight":   "31",
+				"pygmy.purpose":  "showkeys",
+				"pygmy.weight":   "32",
 			},
 		},
 		HostConfig: container.HostConfig{
-			IpcMode:     "private",
 			AutoRemove:  true,
 			VolumesFrom: []string{"amazeeio-ssh-agent"},
 		},
